Add files via upload
Used Appify to scrape across the AllRecipes website. The webscraper scripts were the original attempts to scrape this data.
MergeJSON2 merges the JSON files from the Data folder. 
splittingIngredients.py splits the dataframe on the ingredient column, so each ingredient has a different column
splittingDirections.py splits the splitIngredient dataframe on the directions/method column, so each step has a different column.
The next step is to split the amount of each ingredient and the ingredient into separate columns, so unique ingredients can be individual nodes.